Login works with database
I have made it so that the login menu works with the database, The code
does three checks once the login button is hit. The checks are if the
username and password are on the database within the same row. Then the
final check is the position the user has. Depending on the position it
will open up a different menu.