More database connections
I have added the code to make the card and cheque payment connect to the
database, I still need to figure out how to get the system clock date
and time so that I can pass this into the database to. I have also
retrieved data from the Jobsheet to display to the Mechanic.