2 Huge changes
I have successfully made sure that one button press would insert the
variables into two different tables. This is useful for tables such as
Cheque payment and the payment table itself. Both omgitsmotrix and I
have implemented a session connection to the database, before we could
connect to the database when ever we needed to insert, delete.. etc.